==> code/wrangling_code/3_df_master.qmd <==
---
title: "ENV_COP_DUST"
format:
  html:
    theme: default
---

**This code combines clean data sets (dust, copies, and env)**

# Getting Started

#### Libraries

```{r}
library(tidyverse)
library(here)
```

#### Load Clean Data

```{r}
dust<- readRDS(here("data", "processed_data", "dust_clean.rds")) #Clean dust data
copies<- read_csv(here("data", "processed_data", "total_clean_copies.csv")) #Clean Copies Data
env<- read_csv(here("data", "processed_data", "ysi_2.csv")) #Clean Environmental Data
```

# Further Cleaning

#### Drop NAs from Environmental Data

```{r}
env<- env %>%
  filter(!date == as.Date('2022-02-16') | !site == "RD") %>% #Gets rid of NA
  select(1:20) #Removes NA columns
```

#### Combine 3 data sets to make a large master data set

```{r}
master<- dust %>% full_join(env) %>% full_join(copies)
```

#### Clean Data Again

We will get warnings since we have dust data for the whole year, but Nutrient and Copies_mL data for just the monthly and daily time series. This is ok.

```{r}
master<- master %>%
  select(!...1)  %>% #Gets rid of number column from excel
  mutate(
    sal = as.numeric(master$sal), #Set character columns as numeric
    temp = as.numeric(master$temp),
    do_mgl = as.numeric(master$do_mgl),
    do_per = as.numeric(master$do_per),
    pH = as.numeric(master$pH),
    secchi = as.numeric(master$secchi),
    amm = as.numeric(master$amm),
    nn = as.numeric(master$nn),
    orthop = as.numeric(master$orthop),
    sil = as.numeric(master$sil),
    din_dip = as.numeric(master$din_dip),
    tdn = as.numeric(master$tdn),
    don = as.numeric(master$don),
    doc = as.numeric(master$doc),
    toc = as.numeric(master$toc),
    tn = as.numeric(master$tn)) %>%
    mutate(site = recode(site, "BO" = "Blind Oso", #Rename Site Columns
                             "C2" = "Canals",
                            "RD" = "Gulf"))
```

#### Save as CSV

```{r}
write.csv(master,file= here("data","processed_data","tx_master.csv")) 

```
